* arts-0-0-114
	* vecmat.h: Fixed bug in transf() function. The type of the
	mathematical function is now double (&my_func)(double). Numeric
	does not work here, since mathematical functions for float do not
	exist.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@770 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/arts.h
@@ -88,7 +88,7 @@
 /** The type to use for all floating point numbers. You should never
     use float or double explicitly, unless you have a very good
     reason. Always use this type instead.  */
-typedef double Numeric;
+typedef float Numeric;
 
 //--------------------< Set integer type >--------------------
 /** The type to use for all integer numbers and indices. You should never
